Refactor: api 토큰 수정
