+++ slackmessage/src/main/java/com/ojo/mullyuojo/slackmessage/domain/slackmessage/SlackMessageBotService.java
@@ -1,4 +1,4 @@
-package com.ojo.mullyuojo.slackmessage;
+package com.ojo.mullyuojo.slackmessage.domain.slackmessage;
 
 import com.slack.api.Slack;
 import com.slack.api.methods.MethodsClient;
@@ -7,6 +7,7 @@ import com.slack.api.methods.SlackApiException;
 import com.slack.api.methods.request.chat.ChatPostMessageRequest;
 import com.slack.api.methods.response.chat.ChatPostMessageResponse;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
@@ -15,12 +16,11 @@ import java.io.IOException;
 @Slf4j
 public class SlackMessageBotService {
 
-    private String API_KEY = "xoxb-9346624036995-9369196641863-xXsGA9Hvv0p3gi10HYShClvq";
-    private String CHANNEL_ID = "C09AV6372S3";
-
     Slack slack = Slack.getInstance();
-    String botToken ="xoxb-9346624036995-9369196641863-xXsGA9Hvv0p3gi10HYShClvq";
-    String botToken2 = "xoxb-9380129097203-9385345045154-KVNF8THKQLK8fuPAGjDVgefr"; // my testspace
+
+    @Value("${slack.bot.token}")
+    String botToken2;
+
     MethodsClient methods = slack.methods(botToken2);
 
     public void sendMessage(String message,String recipient){
